Initial commit of binary heap code and test bench
Functions to create and delete binary heaps of ints and insert and
extract int elems from heaps and a test bench to test the
implementation. The functions are built to take heaps and elements and
return status so that NULLs and invariant violations can be caught, e.g.
the heap has been initialized.

ARRAY_SIZE is also implemented, demonstrated and unit tested.

I wrote this to demonstrate good-ish implementation practices. If
something could be better, please let me know.  For instance, I've used
Doxygen to document the functions as opposed to
http://www.sphinx-doc.org/en/stable/domains.html. What are your
thoughts?